refactor(tui): 将 Update 逻辑抽离到 tui_update.go，tui.go 精简为 struct 核心
